refactor register data access classes
refactor anc and family register data access classes
