+++ android/engine/src/main/java/org/smartregister/fhircore/engine/data/local/register/dao/RegisterDaoFactory.kt
@@ -14,13 +14,11 @@
  * limitations under the License.
  */
 
-package org.smartregister.fhircore.engine.data.local.patient.dao.register
+package org.smartregister.fhircore.engine.data.local.register.dao
 
 import javax.inject.Inject
 import javax.inject.Singleton
 import org.smartregister.fhircore.engine.appfeature.model.HealthModule
-import org.smartregister.fhircore.engine.data.local.patient.dao.register.anc.AncPatientRegisterDao
-import org.smartregister.fhircore.engine.data.local.patient.dao.register.family.FamilyRegisterDao
 import org.smartregister.fhircore.engine.domain.repository.RegisterDao
 
 @Singleton
